:recycle: Refactor the addImage method to its own controller class
